requirements.txt
aiogram==3.17.0
aiohttp==3.11.11
python-dotenv==1.0.1
python-stdnum==1.20
fastapi==0.115.7
uvicorn==0.34.0
requests==2.32.3
